Summary of major changes html/css
For HTML, I added more to the content inside (including my selfie and some of my artwork, and even tried having a background audio, which seems like it won't work).
For CSS, I changed the background to be that of cherry blossoms instead of just a single color (its not fun that way). I made the image float on the left side in order to have the text box beside it (instead of below it). I messed with the text/border's margins to make it look geometric, yet appealing and even added a ridged outline (to make it look like its in a frame.